requirements.txt
asgiref==3.8.1
Django==5.0.4
djongo
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
django-redis
django_gridfs_storage
pymongo
Pillow
redis
requests
django-extensions